Check for undefined macros
This change introduces a generalized routine rpmExpandMacros() to expand
macros on behalf of user input, with improved diagnostic facilities.
In particular, one of its arguments is a callback function which is
called whenever an undefined macro is encountered in user input.

When specfile parser calls this routine, it supplies a custom function
to handle undefined macros.  This function is keenly aware of specfile
syntax: valid specfile tokens, such as %prep and %setup, when found in
the right context, are not treated as undefined macros.  Otherwise,
there are 3 possibilities.
- Undefined macros in preamble and %pre/%post/... scriptlets raise
  an error, unless a non-build parse is performed; this is to prevent
  malformed packages from being built.
- In other cases, a warning is issued.
- Some tokens are permitted beyond their context (e.g. %ghost in
  %changelog), to comply with traditional/sloppy usage.  Also, undefined
  macros found in comments normally do not produce a warning.

Update:
- When checking for comments, allow leading whitespace.
- In copyNextLineFromOFI, eliminate rstrlcpy completely: instead of
  truncating the expanded output, simply replace spec->lbuf.