Improve macro table performance
In the existing implementation, when a new macro is added, the whole
table has to be sorted again.  Hence the cost of adding n macros is
worse than O(n^2), due to arithmetic progression.

This change drops all qsort(3) stuff altogether, by carefully preserving
table in sorted order.  In findEntry routine, bsearch(3) is replaced
with customized binary search which tracks position for insertion.
In the addMacro routine, if a matching entry is not found, this
position is used for direct insertion, after the rest of the elements
are "shifted to the right" with memmove(3).  Likewise, in delMacro
routine, the elements are shifted back to the left when the last macro
definition is popped.  Technically, shifting half of the array with
memmove(3) is still O(n^2); however, modern CPUs process contiguous
memory in a very efficient manner, and glibc provides a fine-tuned
memmove(3) implementation.

Also, macro table entries are now allocated in a single chunk.

This change reduces rpm startup costs by factor of 6 (see callgrind
annotations below).  Also, this change improves specfile parser
performance by a factor of 2 (e.g. the parse time of texlive.spec
is reduced from 67s to 35s).

Callgrind annotations for 'rpm --eval %_libdir', previous commit:
88,193,084  PROGRAM TOTALS
27,669,049  glibc-2.16-75f0d304/stdlib/msort.c:msort_with_tmp.part.0'2
24,880,649  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strcmp.S:__GI_strcmp
11,111,276  rpmio/macro.c:compareMacroName
 7,048,036  glibc-2.16-75f0d304/string/../sysdeps/x86_64/memcpy.S:__GI_memcpy
 1,943,118  glibc-2.16-75f0d304/stdlib/msort.c:msort_with_tmp.part.0
 1,903,867  glibc-2.16-75f0d304/elf/dl-lookup.c:do_lookup_x
 1,701,781  rpmio/macro.c:rpmLoadMacroFile
 1,491,535  rpmio/macro.c:doDefine
 1,439,321  rpmio/macro.c:sortMacroTable
 1,254,448  glibc-2.16-75f0d304/malloc/malloc.c:_int_malloc
   683,782  glibc-2.16-75f0d304/malloc/malloc.c:_int_free
   482,748  glibc-2.16-75f0d304/elf/dl-lookup.c:_dl_lookup_symbol_x
   397,935  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strcmp.S:strcmp'2
   307,673  glibc-2.16-75f0d304/elf/dl-misc.c:_dl_name_match_p
   306,247  rpm-4.10.2/lib/rpmrc.c:doReadRC
   300,710  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strcpy.S:__GI_strcpy
   263,309  glibc-2.16-75f0d304/malloc/malloc.c:malloc
   254,657  rpm-4.10.2/lib/rpmrc.c:doReadRC'2
   240,893  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strlen.S:__GI_strlen
...

Callgrind annotations for 'rpm --eval %_libdir', this commit:
14,813,520  PROGRAM TOTALS
1,903,161  glibc-2.16-75f0d304/elf/dl-lookup.c:do_lookup_x
1,701,781  rpmio/macro.c:rpmLoadMacroFile
1,491,535  rpmio/macro.c:doDefine
  932,161  glibc-2.16-75f0d304/malloc/malloc.c:_int_malloc
  900,004  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strcmp.S:__GI_strcmp
  520,864  glibc-2.16-75f0d304/malloc/malloc.c:_int_free
  482,756  glibc-2.16-75f0d304/elf/dl-lookup.c:_dl_lookup_symbol_x
  397,574  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strcmp.S:strcmp'2
  307,322  glibc-2.16-75f0d304/elf/dl-misc.c:_dl_name_match_p
  306,247  rpm-4.10.2/lib/rpmrc.c:doReadRC
  286,734  glibc-2.16-75f0d304/string/../sysdeps/x86_64/memcpy.S:__GI_memcpy
  254,657  rpm-4.10.2/lib/rpmrc.c:doReadRC'2
  240,179  glibc-2.16-75f0d304/string/../sysdeps/x86_64/strlen.S:__GI_strlen
...